requirements.txt
# Core Framework
flask==2.3.3
python-dotenv==1.0.0

# AI/ML Stack
langchain==0.1.20
openai==1.35.0
chromadb==0.4.15
sentence-transformers==2.2.2

# Document Processing
PyPDF2==3.0.1
pypdf==3.15.4

# Data Processing
pandas==2.0.3
numpy==1.24.4
tiktoken==0.5.1

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Testing
pytest==7.4.0
pytest-flask==1.2.0

# CORS for API
flask-cors==4.0.0